@@ -18,15 +18,15 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  amplify_flutter:
-  amplify_datastore:
-  amplify_api:
+  amplify_flutter: ^0.2.4
+  amplify_datastore: ^0.2.4
+  amplify_api: ^0.2.4
 
 
   # The following adds the Cupertino Icons font to your application.
